refactor(circle-progress): move to script setup
