+++ src/packages/__VUE/circleprogress/doc.taro.md
@@ -58,6 +58,18 @@ app.use(CircleProgress);
 | --- | --- |
 | default | 自定义文字内容 |
 
+### 类型定义 version
+
+组件导出以下类型定义：
+
+```js
+import type {
+  CircleProgressStrokeLinecap,
+  CircleProgressProps,
+  CircleProgressInstance
+} from '@nutui/nutui-taro';
+```
+
 ## 主题定制
 
 ### 样式变量
